% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemmtan-package.R
\docType{package}
\name{riemmtan-package}
\alias{riemmtan}
\alias{riemmtan-package}
\title{riemmtan: Riemannian Metrics for Symmetric Positive Definite Matrices}
\description{
Implements various Riemannian metrics for symmetric positive definite matrices, including AIRM (Affine Invariant Riemannian Metric), Log-Euclidean, Euclidean, Log-Cholesky, and Bures-Wasserstein metrics. Provides functions for computing logarithmic and exponential maps, vectorization, and statistical operations on the manifold of positive definite matrices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nicoesve.github.io/riemmtan/}
  \item Report bugs at \url{https://github.com/nicoesve/riemmtan/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Escobar \email{nescoba@iu.edu} (\href{https://orcid.org/0009-0006-0800-5692}{ORCID})

Other contributors:
\itemize{
  \item Jaroslaw Harezlak [thesis advisor]
}

}
\keyword{internal}
